centralize setup of Deno namespace 1/2
