+++ cli/js/runtime.ts
@@ -1,14 +1,23 @@
 // Copyright 2018-2020 the Deno authors. All rights reserved. MIT license.
 import { core } from "./core.ts";
+import * as Deno from "./deno.ts";
 import * as dispatchMinimal from "./ops/dispatch_minimal.ts";
 import * as dispatchJson from "./ops/dispatch_json.ts";
 import { assert } from "./util.ts";
 import * as util from "./util.ts";
 import { setBuildInfo } from "./build.ts";
-import { setVersions } from "./version.ts";
 import { LocationImpl } from "./web/location.ts";
 import { setPrepareStackTrace } from "./error_stack.ts";
 import { Start, start as startOp } from "./ops/runtime.ts";
+import { setSignals } from "./process.ts";
+import { symbols } from "./symbols.ts";
+import { internalObject } from "./internals.ts";
+
+interface Version {
+  deno: string;
+  v8: string;
+  typescript: string;
+}
 
 export let OPS_CACHE: { [name: string]: number };
 
@@ -43,26 +52,34 @@ export function start(preserveDenoNamespace = true, source?: string): Start {
   // args and other info.
   const s = startOp();
 
-  setVersions(s.denoVersion, s.v8Version, s.tsVersion);
+  // Add internal object to Deno object.
+  // This is not exposed as part of the Deno types.
+  // @ts-ignore
+  Deno[symbols.internal] = internalObject;
+  // Build info is used by internal code, so setting it first.
   setBuildInfo(s.os, s.arch);
+
   util.setLogDebug(s.debugFlag, source);
   util.immutableDefine(globalThis, "location", new LocationImpl(s.location));
   Object.freeze(globalThis.location);
-
   setPrepareStackTrace(Error);
-
-  // TODO(bartlomieju): I don't like that it's mixed in here, when
-  // compiler and worker runtimes call this funtion and they don't use
-  // Deno namespace (sans shared queue - Deno.core)
-
-  // pid and noColor need to be set in the Deno module before it's set to be
-  // frozen.
-  util.immutableDefine(globalThis.Deno, "pid", s.pid);
-  util.immutableDefine(globalThis.Deno, "noColor", s.noColor);
-  Object.freeze(globalThis.Deno);
+  setSignals();
 
   if (preserveDenoNamespace) {
-    util.immutableDefine(globalThis, "Deno", globalThis.Deno);
+    const version: Version = {
+      deno: s.denoVersion,
+      v8: s.v8Version,
+      typescript: s.tsVersion
+    };
+    Object.freeze(version);
+    util.immutableDefine(Deno, "version", version);
+    util.immutableDefine(Deno, "pid", s.pid);
+    util.immutableDefine(Deno, "noColor", s.noColor);
+    util.immutableDefine(Deno, "args", [...s.args]);
+    // TODO(bartlomieju): Object.freeze should be called recursively on
+    // all properties of `globalThis.Deno`
+    Object.freeze(Deno);
+    util.immutableDefine(globalThis, "Deno", Deno);
     // Deno.core could ONLY be safely frozen here (not in globals.ts)
     // since shared_queue.js will modify core properties.
     Object.freeze(globalThis.Deno.core);
@@ -74,5 +91,7 @@ export function start(preserveDenoNamespace = true, source?: string): Start {
     assert(globalThis.Deno === undefined);
   }
 
+  util.log("cwd", s.cwd);
+  util.log("args", Deno.args);
   return s;
 }
