+++ cli/js/version.ts
@@ -1,9 +1,4 @@
 // Copyright 2018-2020 the Deno authors. All rights reserved. MIT license.
-interface Version {
-  deno: string;
-  v8: string;
-  typescript: string;
-}
 
 export const version: Version = {
   deno: "",
